feat: compress slash command
