+++ .deploy/webapp/package.json
@@ -73,7 +73,7 @@
     "@sentry/tracing": "^7.38.0",
     "@swc/helpers": "0.5.3",
     "@tinymce/tinymce-angular": "^6.0.1",
-    "@xpert-ai/chatkit-ui": "~0.3.6",
+    "@xpert-ai/chatkit-ui": "~0.3.8",
     "angular-gridster2": "^14.0.1",
     "apache-arrow": "^9.0.0",
     "axios": "1.6.8",
@@ -204,7 +204,7 @@
   "pnpm": {
     "overrides": {
       "@langchain/core": "0.3.72",
-      "@xpert-ai/chatkit-web-component": "0.3.3",
+      "@xpert-ai/chatkit-web-component": "~0.3.4",
       "langchain": "0.3.31",
       "zod": "3.25.67",
       "minimatch": "9.0.4"
