+++ apps/cloud/src/app/xpert/ai-message/content/content.component.ts
@@ -1,11 +1,12 @@
 import { toObservable } from '@angular/core/rxjs-interop'
 import { ChangeDetectionStrategy, Component, computed, inject, input, signal } from '@angular/core'
-import { DateRelativePipe } from '@cloud/app/@core'
+import { CONTEXT_COMPRESSION_COMPONENT_TYPE, DateRelativePipe } from '@cloud/app/@core'
 import { TMessageContentComplex, TMessageContentText } from '@cloud/app/@core/types'
 import { Copy2Component } from '@cloud/app/@shared/common'
 import { listEnterAnimation } from '@xpert-ai/core'
 import { TranslateModule } from '@ngx-translate/core'
 import { MarkdownModule } from 'ngx-markdown'
+import { ChatContextCompressionChunkComponent } from '@cloud/app/@shared/chat/context-compression/context-compression.component'
 import { ChatService } from '../../chat.service'
 import { ChatComponentMessageComponent } from '../../component-message'
 import { TCopilotChatMessage } from '../../types'
@@ -20,6 +21,7 @@ import { splitStreamingMarkdown } from './streaming-markdown'
     ...ZardTooltipImports,
     DateRelativePipe,
     Copy2Component,
+    ChatContextCompressionChunkComponent,
     ChatComponentMessageComponent,
     ChatMessageDashboardComponent
   ],
@@ -37,6 +39,7 @@ export class ChatMessageContentComponent {
   readonly last = input<boolean>()
   readonly collapse = input<boolean>()
   readonly agentOutput = input<boolean>(false)
+  readonly contextCompressionComponentType = CONTEXT_COMPRESSION_COMPONENT_TYPE
 
   readonly submessage = computed(() => this.content())
   readonly text = computed(() => (this.content()?.type === 'text' ? (<TMessageContentText>this.content()).text : ''))
