+++ apps/cloud/src/assets/i18n/en.json
@@ -610,7 +610,10 @@
       "CommandUnavailable": "Unavailable",
       "PreviewStopGenerate": "Closing the chat panel will stop generation.",
       "StopGenerateOnRestart": "Restarting the conversation will stop the current generation.",
-      "ContextCompression": "Context compression...",
+      "ContextCompression": "Automatically compressing context",
+      "ContextCompressed": "Context automatically compressed",
+      "ContextCompressionSkipped": "Context not compressed",
+      "ContextCompressionFailed": "Context compression failed",
       "Event_on_agent_start": "Agent start",
       "Event_on_agent_end": "Agent end",
       "Event_on_tool_start": "Tool start",
