+++ apps/cloud/src/assets/i18n/zh-CN.json
@@ -278,7 +278,10 @@
     "Chat": {
       "QuoteSelection": "引用选中内容",
       "RemoveReference": "移除引用",
-      "ContextCompression": "上下文压缩..."
+      "ContextCompression": "正在自动压缩上下文",
+      "ContextCompressed": "上下文已自动压缩",
+      "ContextCompressionSkipped": "上下文无需压缩",
+      "ContextCompressionFailed": "上下文压缩失败"
     },
     "MENU": {
       "Certification": "认证",
