+++ apps/cloud/src/assets/i18n/zh-Hans.json
@@ -2946,7 +2946,10 @@
     "Chat": {
       "QuoteSelection": "引用选中内容",
       "RemoveReference": "移除引用",
-      "ContextCompression": "上下文压缩...",
+      "ContextCompression": "正在自动压缩上下文",
+      "ContextCompressed": "上下文已自动压缩",
+      "ContextCompressionSkipped": "上下文无需压缩",
+      "ContextCompressionFailed": "上下文压缩失败",
       "SearchChats": "搜索聊天...",
       "Input": "输入",
       "Output": "输出",
@@ -2981,6 +2984,7 @@
       "RetrievingDocuments": "检索文档",
       "RetrievedDocuments": "检索到{{value}}个相关文档块",
       "CommonRoleDescription": "嗨，我能帮什么忙吗？我可以聊天，检索知识库。想要使用工具请选择相应的角色。",
+      "CommonWelcomeDescription": "从工作区选择一位数字专家，或直接进入通用聊天。",
       "CloseSidebar": "关闭边栏",
       "OpenSidebar": "打开边栏",
       "NewChat": "新聊天",
