+++ packages/contracts/src/index.ts
@@ -1,4 +1,6 @@
 export * from './types'
+export type { I18nObject, I18nText } from './i18n.model'
+export { resolveI18nText } from './i18n.model'
 export * from './contact.model'
 export * from './core.model'
 export * from './country.model'
