+++ packages/plugins/agent-middlewares/src/lib/middlewares.module.ts
@@ -1,51 +1,49 @@
-import { XpertServerPlugin, IOnPluginBootstrap, IOnPluginDestroy } from '@xpert-ai/plugin-sdk';
-import { CqrsModule } from '@nestjs/cqrs';
-import chalk from 'chalk';
-import { TodoListMiddleware } from './todoListMiddleware';
-import { SummarizationMiddleware } from './summarization';
-import { ClientEffectMiddleware } from './clientEffect';
-import { LLMToolSelectorNameMiddleware } from './llmToolSelector';
-import { LocalShellSandboxProvider } from './localShellSandbox.provider';
-import { StructuredOutputMiddleware } from './structuredOutput';
-import { SandboxFileMiddleware } from './sandboxFile';
-import { ContextCompressionMiddleware } from './contextCompression';
+import { XpertServerPlugin, IOnPluginBootstrap, IOnPluginDestroy } from '@xpert-ai/plugin-sdk'
+import { CqrsModule } from '@nestjs/cqrs'
+import chalk from 'chalk'
+import { TodoListMiddleware } from './todoListMiddleware'
+import { SummarizationMiddleware } from './summarization'
+import { ClientEffectMiddleware } from './clientEffect'
+import { LLMToolSelectorNameMiddleware } from './llmToolSelector'
+import { LocalShellSandboxProvider } from './localShellSandbox.provider'
+import { StructuredOutputMiddleware } from './structuredOutput'
+import { SandboxFileMiddleware } from './sandboxFile'
 
 @XpertServerPlugin({
-	/**
-	 * An array of modules that will be imported and registered with the plugin.
-	 */
-	imports: [CqrsModule],
+  /**
+   * An array of modules that will be imported and registered with the plugin.
+   */
+  imports: [CqrsModule],
 
-    providers: [
-		SummarizationMiddleware,
-        TodoListMiddleware,
-		LLMToolSelectorNameMiddleware,
-		ClientEffectMiddleware,
-		SandboxFileMiddleware,
-		LocalShellSandboxProvider,
-		StructuredOutputMiddleware,
-		ContextCompressionMiddleware
-    ],
+  providers: [
+    SummarizationMiddleware,
+    TodoListMiddleware,
+    LLMToolSelectorNameMiddleware,
+    ClientEffectMiddleware,
+    SandboxFileMiddleware,
+    LocalShellSandboxProvider,
+    StructuredOutputMiddleware
+  ]
 })
 export class AgentMiddlewaresModule implements IOnPluginBootstrap, IOnPluginDestroy {
-	// We disable by default additional logging for each event to avoid cluttering the logs
-	private logEnabled = true;
+  // We disable by default additional logging for each event to avoid cluttering the logs
+  private logEnabled = true
 
-	/**
-	 * Called when the plugin is being initialized.
-	 */
-	onPluginBootstrap(): void | Promise<void> {
-		if (this.logEnabled) {
-			console.log(chalk.green(`${AgentMiddlewaresModule.name} is being bootstrapped...`));
-		}
-	}
+  /**
+   * Called when the plugin is being initialized.
+   */
+  onPluginBootstrap(): void | Promise<void> {
+    if (this.logEnabled) {
+      console.log(chalk.green(`${AgentMiddlewaresModule.name} is being bootstrapped...`))
+    }
+  }
 
-	/**
-	 * Called when the plugin is being destroyed.
-	 */
-	onPluginDestroy(): void | Promise<void> {
-		if (this.logEnabled) {
-			console.log(chalk.green(`${AgentMiddlewaresModule.name} is being destroyed...`));
-		}
-	}
+  /**
+   * Called when the plugin is being destroyed.
+   */
+  onPluginDestroy(): void | Promise<void> {
+    if (this.logEnabled) {
+      console.log(chalk.green(`${AgentMiddlewaresModule.name} is being destroyed...`))
+    }
+  }
 }
