+++ packages/server-ai/src/ai/runtime-command.service.spec.ts
@@ -1,3 +1,5 @@
+import { LanguagesEnum } from '@xpert-ai/contracts'
+import { RequestContext } from '@xpert-ai/server-core'
 import { RuntimeCommandService } from './runtime-command.service'
 import type { RuntimeSlashCommand } from './runtime-command.guards'
 
@@ -333,8 +335,26 @@ describe('RuntimeCommandService', () => {
                         description: 'Run a goal loop',
                         argsHint: '<objective>',
                         action: {
-                            type: 'submit_prompt',
+                            type: 'insert_invocation',
                             template: 'Goal: {{args}}'
+                        }
+                    },
+                    {
+                        name: 'compact',
+                        label: {
+                            en_US: 'Compress',
+                            zh_Hans: '压缩'
+                        },
+                        description: {
+                            en_US: 'Compress this thread context',
+                            zh_Hans: '压缩此线程的上下文'
+                        },
+                        category: 'session',
+                        kind: 'command',
+                        aliases: ['compress'],
+                        action: {
+                            type: 'submit_prompt',
+                            template: '/compact'
                         }
                     },
                     {
@@ -389,8 +409,85 @@ describe('RuntimeCommandService', () => {
                     nodeKey: 'middleware-goal',
                     label: 'Goal Loop'
                 }
+            }),
+            expect.objectContaining({
+                name: 'compact',
+                label: 'Compress',
+                description: 'Compress this thread context',
+                category: 'session',
+                aliases: ['compress'],
+                action: {
+                    type: 'submit_prompt',
+                    template: '/compact',
+                    runtimeCapabilities: {
+                        mode: 'allowlist',
+                        skills: {
+                            ids: []
+                        },
+                        plugins: {
+                            nodeKeys: ['middleware-goal']
+                        },
+                        subAgents: {
+                            nodeKeys: []
+                        }
+                    }
+                },
+                source: {
+                    type: 'middleware',
+                    provider: 'goal-loop-provider',
+                    nodeKey: 'middleware-goal',
+                    label: 'Goal Loop'
+                }
             })
         ])
+    })
+
+    it('localizes middleware runtime command text from i18n objects', () => {
+        const languageSpy = jest
+            .spyOn(RequestContext, 'getLanguageCode')
+            .mockReturnValue(LanguagesEnum.SimplifiedChinese)
+        const service = new RuntimeCommandService()
+
+        try {
+            expect(
+                service.normalizeMiddlewareRuntimeSlashCommands(
+                    [
+                        {
+                            name: 'compact',
+                            label: {
+                                en_US: 'Compress',
+                                zh_Hans: '压缩'
+                            },
+                            description: {
+                                en_US: 'Compress this thread context',
+                                zh_Hans: '压缩此线程的上下文'
+                            },
+                            action: {
+                                type: 'submit_prompt',
+                                template: '/compact'
+                            }
+                        }
+                    ],
+                    {
+                        provider: 'ContextCompressionMiddleware',
+                        nodeKey: 'middleware-compression',
+                        label: 'Context Compression'
+                    }
+                )
+            ).toEqual([
+                expect.objectContaining({
+                    name: 'compact',
+                    label: '压缩',
+                    description: '压缩此线程的上下文',
+                    workflow: expect.objectContaining({
+                        label: '压缩',
+                        description: '压缩此线程的上下文'
+                    })
+                })
+            ])
+        } finally {
+            languageSpy.mockRestore()
+        }
     })
 
     it('merges runtime commands by middleware, xpert, preferred skill, workspace, then skill priority', () => {
