+++ packages/server-ai/src/xpert-middleware/index.ts
@@ -1,5 +1,6 @@
 export * from './browser-automation.middleware'
 export * from './client-tool.middleware'
+export * from './context-compression.middleware'
 export * from './human-in-the-loop.middleware'
 export * from './ralph-loop.middleware'
 export * from './xpert-middleware.module'
