+++ packages/server-ai/src/xpert-middleware/xpert-middleware.module.ts
@@ -1,11 +1,24 @@
 import { Module } from '@nestjs/common'
 import { BrowserAutomationMiddleware } from './browser-automation.middleware'
 import { ClientToolMiddleware } from './client-tool.middleware'
+import { ContextCompressionMiddleware } from './context-compression.middleware'
 import { HumanInTheLoopMiddleware } from './human-in-the-loop.middleware'
 import { RalphLoopMiddleware } from './ralph-loop.middleware'
 
 @Module({
-    providers: [BrowserAutomationMiddleware, ClientToolMiddleware, HumanInTheLoopMiddleware, RalphLoopMiddleware],
-    exports: [BrowserAutomationMiddleware, ClientToolMiddleware, HumanInTheLoopMiddleware, RalphLoopMiddleware]
+    providers: [
+        BrowserAutomationMiddleware,
+        ClientToolMiddleware,
+        ContextCompressionMiddleware,
+        HumanInTheLoopMiddleware,
+        RalphLoopMiddleware
+    ],
+    exports: [
+        BrowserAutomationMiddleware,
+        ClientToolMiddleware,
+        ContextCompressionMiddleware,
+        HumanInTheLoopMiddleware,
+        RalphLoopMiddleware
+    ]
 })
 export class XpertMiddlewareModule {}
